Change some function for better flow
