+++ app/views/DR/HomeScreen/index.js
@@ -1,7 +1,7 @@
 import 'moment/locale/es';
 
 import moment from 'moment';
-import { Button, Card, Left, Text } from 'native-base';
+import { Button, Left, Text } from 'native-base';
 import React, { Component } from 'react';
 import { withTranslation } from 'react-i18next';
 import {
@@ -29,7 +29,7 @@ import fetch from '../../../helpers/Fetch';
 import { GetStoreData, SetStoreData } from '../../../helpers/General';
 import { getAllCases } from '../../../services/DR/getAllCases.js';
 import DialogAdvice from '../../DialogAdvices';
-import Index from '../Sponsors';
+import DashboardCards from './DashboardCards';
 import styles from './style';
 
 class HomeScreen extends Component {
@@ -165,46 +165,6 @@ class HomeScreen extends Component {
       state: { confirmed, deaths, recovered, current, refreshing, notified },
     } = this;
 
-    const dashboardCards = () => {
-      const dashboardCategories = [
-        { label: 'label.positive_label', styles: '', value: confirmed },
-        {
-          label: 'label.deceased_label',
-          styles: { color: Colors.BUTTON_LIGHT_TEX },
-          value: deaths,
-        },
-        {
-          label: 'label.recovered_label',
-          styles: { color: Colors.GREEN },
-          value: recovered,
-        },
-        {
-          label: 'label.case_day_label',
-          styles: { color: Colors.SUN },
-          value: current,
-        },
-      ];
-      const CONFIRMED_COVID_CASES_URL =
-        'https://coronavirusrd.gob.do/casos-de-covid-19-confirmados/';
-
-      return dashboardCategories.map(obj => 
-         (
-          <TouchableOpacity
-            style={styles.shadowBorder}
-            key={Index}
-            onPress={() =>
-              navigation.navigate('Details', {
-                source: { uri: CONFIRMED_COVID_CASES_URL },
-              })
-            }>
-            <Card style={styles.infoCards}>
-              <Text style={[styles.dataText, obj.styles]}>{obj.value}</Text>
-              <Text style={styles.text}>{t(obj.label)}</Text>
-            </Card>
-          </TouchableOpacity>
-        );
-      });
-    };
     return (
       <SafeAreaView style={{ flex: 1 }}>
         <View style={{ flex: 1, backgroundColor: Colors.BLUE_RIBBON }}>
@@ -250,7 +210,14 @@ class HomeScreen extends Component {
                     </View>
                   </View>
                   <View style={styles.actualSituationContainer}>
-                    {dashboardCards()}
+                    <DashboardCards
+                      t={t}
+                      navigation={navigation}
+                      confirmed={confirmed}
+                      deaths={deaths}
+                      recovered={recovered}
+                      current={current}
+                    />
                   </View>
                   <LocationMatch navigation={this.props.navigation} />
                   <Aurora navigation={this.props.navigation} />
